requirements.txt
numpy
pygame
opencv_python
PyQt6
openai
opencc-python-reimplemented
pydantic
pandas
requests
onnxruntime
rembg
google-genai
tiktoken
pyaudio
anthropic
PySide6
pyyaml
vosk
gpu-list
# 开发时 TTS 整合包 7z 解压（可选，也可用 build_exe/7za.exe）
py7zr